feat(graph layout): align branches in timelines
