+++ src/graph/types.ts
@@ -1,0 +1,3 @@
+export interface TimeNode {
+    timestamp: number | undefined;
+}
